+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentSheetFlowControllerViewController.swift
@@ -132,10 +132,6 @@ class PaymentSheetFlowControllerViewController: UIViewController {
             walletOptions.insert(.link)
         }
 
-        if !intent.isPaymentIntent {
-            walletOptions.insert(.setupIntent)
-        }
-
         let header = WalletHeaderView(
             options: walletOptions,
             appearance: configuration.appearance,