+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentSheetViewController.swift
@@ -114,10 +114,6 @@ class PaymentSheetViewController: UIViewController {
             walletOptions.insert(.link)
         }
 
-        if !intent.isPaymentIntent {
-            walletOptions.insert(.setupIntent)
-        }
-
         let header = WalletHeaderView(
             options: walletOptions,
             appearance: configuration.appearance,